Initial commit
Currently an output of all TRUE just means that the name _can_ be
written using some combination of the symbols for the elements of the
periodic table.  Getting the output for that should be pretty straight
forward, but I'm getting to the point where I'm quite hungry and the
kitchen needs to be cleaned. So I should go do that.

Shout out to the periodic-table JSON file that was put together by
Bowsernator.  That's a cool thing, and it'll make it fun to try and come
up with a visualization for the output of this in a web application or
something.