Add javascript version
Getting used to the API for the node env.  I've never really scripted in
it and the immediate difficulty is the asynchonous code.  Very cool!
But it's going to force my hand on understanding promises and stuff.
Which is good, I feel like they're not _that_ hard, I just need to
**use** them.